Change coin variable type from integer to double

+++ 2_11.cpp
@@ -4,7 +4,7 @@
 
 int main() {
     // Create variables for each type of coin
-    int pennies, nickels, dimes, quarters, halfDollars;
+    double pennies, nickels, dimes, quarters, halfDollars;
 
     // Ask and store pennies
     std::cout << "How many pennies do you have?\n";
